Added setting of observed data covariance
When the observed data covariance, it is set to an arbitrary value
(1.0) to avoid an infinite log likelihood. This is of the same
magnitude as the observed data in this model, so the prior likelihood
for the parameters was by far the dominant term in the likelihood
calculation, leading to undesired results in the MCMC sampling. Fixed
this by setting the covariance of the observed data very low
(0.01). Now, results from the MCMC are what we would expect from this
test model.