Added method that performs the division of an parameter space dimension
It can either subdivide the dimension evenly or by percentile
according to the parameter's prior distribution. If the prior is a
Gaussian distribution and even dimension subdivision is chosen, then
the parameter range is specified by some number of standard deviations
about the mean where the number of standard deviations is settable by
the user of the class.

In addition, changed the signature of the Generate() method to take a
vector of Parameters rather than two vectors of doubles that specify
the minima and maxima of the parameters.