Fixed problem with trying to load extra files in model_output directory
When checking whether a file in the model_output directory started
with "run", the method std::string::find_first_of() was being
used. This method takes a string, but it does not find the first
location of that string. Rather, it finds the first location of any
characters in that string. Hence, because the string passed to it was
"run", it matched the extra file "results.dat" that was in the model_output
directory that should have been ignored.

Switching to std::string::substr() to check if the directory name
starts with "run" fixes this problem.