Changed argument format for specifying CSV file
madai_generate_mcmc_trace requires only a name of a CSV file that will
be stored in the STATDIR/traces. I made madai_analyze_trace's second
argument follow that format (file name instead of full path). This
will probably be better for ensuring that the parameters listed in the
CSV file match those in the parameter priors file. If there is a valid
use case for the trace file being somewhere else, we can swtich it
back.

Also:

- Updated the manual to reflect the changes mentioned above.

- Adjusted capitalization of some variables to be camelCase.

- Removed splitComma() function and replaced it with existing
  madai::SplitString() function.