Added option SAMPLER_INACTIVE_PARAMETERS_FILE to madai_generate_trace
This file specifies a list of parameters that should be deactivated in
the sampler. The format is

<parameter name> <parameter value>

Note that <paramever value> should be a value that has relatively high
likelihood as determined by the parameter prior. For example, a
parameter value outside the specified range of a uniform distribution
will drastically change the relative likelihood for samples taken by a
sampler.