Use \path{} defined by package hyperref
This version of \path{} has the advantage that it will break a long
path across lines whereas the previous definition that used \texttt{}
did not. It also does not require underscores to be preceded by a
backslash (so a lot of changes had to be made around the documentation
to remove these).